+++ plugins/org.eclipse.app4mc.multicore.openmapping/src/org/eclipse/app4mc/multicore/openmapping/algorithms/ilp/energyminimization/EnergyMinimization.java
@@ -956,7 +956,7 @@ public class EnergyMinimization extends AbstractILPBasedMappingAlgorithm {
 				default:
 					break;
 			}
-			return mul * t.getValue();
+			return mul * t.getValue().longValue();
 		}
 
 		static long nanoTime() {
